office

+++ src/leetcode/a300/_367Solution.java
@@ -1,7 +1,14 @@
 package leetcode.a300;
-
+/**
+ * 判断一个数是不是完全平方数
+ */
 public class _367Solution {
 
+    /**
+     * 二分法快速定位
+     * @param num
+     * @return
+     */
     public boolean isPerfectSquare(int num) {
 
         long tar = (long) num;
@@ -9,7 +16,8 @@ public class _367Solution {
         int r = num / 2 + 1;
         while (l < r) {
             int mid = (l + r) / 2;
-            if (((long)mid *(long)mid) < tar) {
+            // 注意这个地方整数强制转换之后再计算
+            if ((long)mid *mid < tar) {
                 l = mid + 1;
             } else {
                 r = mid;
